Outer loop refactoring
