+++ src/main/java/com/powsybl/openloadflow/OuterLoopConfig.java
@@ -8,7 +8,7 @@ package com.powsybl.openloadflow;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.loadflow.LoadFlowParameters;
-import com.powsybl.openloadflow.ac.OuterLoop;
+import com.powsybl.openloadflow.lf.outerloop.OuterLoop;
 import org.apache.commons.compress.utils.Lists;
 
 import java.util.List;
