+++ src/main/java/com/powsybl/openloadflow/ac/AcLoadFlowParameters.java
@@ -10,6 +10,7 @@ import com.powsybl.math.matrix.MatrixFactory;
 import com.powsybl.openloadflow.lf.AbstractLoadFlowParameters;
 import com.powsybl.openloadflow.ac.equations.AcEquationSystemCreationParameters;
 import com.powsybl.openloadflow.ac.nr.NewtonRaphsonParameters;
+import com.powsybl.openloadflow.lf.outerloop.OuterLoop;
 import com.powsybl.openloadflow.network.LfNetworkParameters;
 import com.powsybl.openloadflow.network.util.VoltageInitializer;
