+++ src/main/java/com/powsybl/openloadflow/ac/OuterLoopContextImpl.java
@@ -7,6 +7,7 @@
 package com.powsybl.openloadflow.ac;
 
 import com.powsybl.openloadflow.ac.nr.NewtonRaphsonResult;
+import com.powsybl.openloadflow.lf.outerloop.OuterLoopContext;
 import com.powsybl.openloadflow.network.LfNetwork;
 
 import java.util.Objects;
